% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dogpirt_responses.R
\name{format_dogpirt_responses}
\alias{format_dogpirt_responses}
\title{Format Responses for the Dynamic, Ordinal GPIRT Model}
\usage{
format_dogpirt_responses(
  dataframe,
  respondent_ids,
  item_ids,
  period_ids,
  responses
)
}
\arguments{
\item{dataframe}{A data.frame containing observed responses}

\item{respondent_ids}{The column indicating which respondent an observation
(i.e. dataframe row) is for}

\item{item_ids}{The column indicating which item an observation
(i.e. dataframe row) is for}

\item{period_ids}{The column indicating which time period an observation
(i.e. dataframe row) is for}

\item{responses}{The column recording the observed response}
}
\value{
A list with the same number of unique elements as
\code{dataframe$period_ids}, where each element is a matrix of the observed
responses in a given time period. The rows of each matrix correspond to a
respondent, the columns of each matrix correspond to an item, and the
elements of the list correspond to a time period. The elments of the list
will be named by the time period IDs, the rows of each matrix will be
named by the respondent IDs, and the columns of each matrix will be named
by the item IDs.
}
\description{
Turn a dataframe of responses from multiple time periods into a list of
matrices for \code{\link[=gpirtMCMC]{gpirtMCMC()}}
}
