% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpirtMCMC.R
\name{gpirtMCMC}
\alias{gpirtMCMC}
\title{Gaussian Process IRT MCMC}
\usage{
gpirtMCMC(responses, sample_iterations, burn_iterations, yea_code = 1,
  nay_code = 0, sf = 1, ell = 1)
}
\arguments{
\item{responses}{A matrix of responses}

\item{sample_iterations}{An integer vector of length one giving the number
of samples to record}

\item{burn_iterations}{An integer vector of length one giving the number of
burn in (unrecorded) iterations}

\item{yea_code}{A vector of length one giving the value corresponding to
a "yea" response; default is 1}

\item{nay_code}{A vector of length one giving the value corresponding to
a "nay" response; default is 0}

\item{sf}{A numeric vector of length one giving the scale factor for the
covariance function for the Gaussian process prior; default is 1}

\item{ell}{A numeric vector of length one giving the length scale for the
covariance function for the Gaussian process prior; default is 1}
}
\value{
A list of length two; the first element, "theta", is a matrix of
  dimensions sample_iterations x n giving the theta parameter draws, and the
  second element, "f", is an array of dimensions n x m x sample_iterations
  giving the f(theta) parameter draws.
}
\description{
Provides posterior samples for the GP-IRT.
}
