% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_irf.R
\name{plot_irf}
\alias{plot_irf}
\title{Plot Item Response Functions}
\usage{
plot_irf(gp_results, theta_star, main_title = "", rug = FALSE,
  responses = NULL, estimates = NULL)
}
\arguments{
\item{gp_results}{A list returned from \code{\link[gpmlr]{gp}}
(see details and examples).}

\item{theta_star}{A numeric vector of theta values to predict the
probability of a 1 response at.}

\item{main_title}{(Optional) A title for the plot}

\item{rug}{A logical vector of length one; if \code{TRUE} a rug of the
theta estimates whose responses were 1 will be placed at the top of
the plot and a rug of the theta estimates for respondents whose
responses were -1 will be placed at the bottom of the plot
(default is \code{FALSE})}

\item{responses}{An integer vector of observed responses;
must be supplied if \code{rug} is \code{TRUE}}

\item{estimates}{A numeric vector of theta estimates for your respondents;
must be supplied if \code{rug} is \code{TRUE}}
}
\description{
Plot Item Response Functions
}
\examples{
\donttest{
    library(gpmlr)
    set.seed(123)
    theta <- rnorm(100)
    f <- 3 * sin(theta)
    P <- 1 / (1 + exp(-f))
    responses <- sapply(P, function(p) sample(c(1, -1), 1, prob = c(p, 1-p)))
    theta_star <- seq(-3, 3, 0.01)
    hyperparameters <- list(mean = numeric(), cov = c(0, 0))
    gp_results <- gp(hyperparameters, "infLaplace", "meanZero",
                     "covSEiso", "likLogistic", theta, responses,
                     theta_star)
    plot_irf(gp_results, theta_star, rug = TRUE, responses = responses,
             estimates = theta)
    z <- 3 * sin(theta_star)
    p <- 1 / (1 + exp(-z))
    lines(theta_star, p, lty = 2)
}

}
