% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_fstar.R
\name{recover_fstar}
\alias{recover_fstar}
\title{Gaussian Process Recover IRT}
\usage{
recover_fstar(
  seed_state,
  f,
  theta,
  beta_prior_means = matrix(0, nrow = 2, ncol = ncol(f)),
  beta_prior_sds = matrix(3, nrow = 2, ncol = ncol(f))
)
}
\arguments{
\item{seed_state}{An integer vector of length 625.}

\item{f}{A matrix of latent function values of shape n by m by h, where
n is number of respondents, m is number of items per session, and
h is number of sessions.}

\item{theta}{A matrix of shape n by h, giving initial values
for the respondent ideology parameters}

\item{beta_prior_means}{A numeric matrix of with \code{ncol(data)} columns
and two rows giving the prior means for the items' linear means' intercept
and slope; by default, a matrix of zeros}

\item{beta_prior_sds}{A numeric matrix of with \code{ncol(data)} columns
and two rows giving the prior standard deviations for the items' linear
means' intercept and slope; by default, a matrix of threes}
}
\value{
A list with elements
  \describe{
      \item{fstar}{Estimated item response functions for the items, with one
                  column per item, or \code{m} columns, and 1001 rows.
                  The first row has the probabilities of a 1 response
                  for a theta value of -5.0, the second the probability for
                  each item of a 1 response for a theta value of -4.99,
                  ..., and the last for a theta value of 5.0.}
  }
}
\description{
Recover fstar from f and seed
}
