% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gpirtMCMC}
\alias{gpirtMCMC}
\title{Gaussian Process IRT MCMC}
\usage{
gpirtMCMC(y, sample_iterations, burn_iterations)
}
\arguments{
\item{y}{A matrix of responses; each entry should be -1 or 1}

\item{sample_iterations}{An integer vector of length one giving the number
of samples to record}

\item{burn_iterations}{An integer vector of length one giving the number of
burn in (unrecorded) iterations}
}
\value{
A list of length two; the first element, "theta", is a matrix of
  dimensions sample_iterations x n giving the theta parameter draws, and the
  second element, "f", is an array of dimensions n x m x sample_iterations
  giving the f(theta) parameter draws.
}
\description{
Provides posterior samples for the GP-IRT.
}
