% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrix.R
\name{response_matrix}
\alias{response_matrix}
\alias{is.response_matrix}
\alias{as.response_matrix}
\title{Create an object of class response_matrix}
\usage{
response_matrix(data, yea_codes = 1, nay_codes = 0,
  missing_codes = NA)

is.response_matrix(x)

as.response_matrix(x, yea_codes = 1, nay_codes = 0,
  missing_codes = NA)
}
\arguments{
\item{data}{A matrix or dataframe of responses}

\item{yea_codes}{A vector giving the values corresponding to a "yea"
response (default is 1)}

\item{nay_codes}{A vector giving the values corresponding to a "nay"
response (default is 0)}

\item{missing_codes}{A vector giving the values corresponding to a missing
response (default is NA)}

\item{x}{An R object}
}
\description{
\code{response_matrix} creates a response_matrix object, for use with
\code{\link{gpirtMCMC}}; \code{is.response_matrix} tests if its argument is
a \code{response_matrix} object; \code{as.response_matrix} attemps to turn
its argument into a \code{response_matrix}.
}
\examples{
## Data in {0, 1}
x   <- c(1, 0, 1, 1, 0, NA)
ex1 <- matrix(x, nrow = 3)
ex2 <- data.frame(x1 = x[1:3], x2 = x[4:6])
response_matrix(ex1)
response_matrix(ex2)
## Multiple "yea" codes
x   <- c(1, -1, 2, 3, -1, NA)
ex3 <- matrix(x, nrow = 3)
ex4 <- data.frame(x1 = x[1:3], x2 = x[4:6])
response_matrix(ex3, yea_codes = 1:3)
response_matrix(ex4, yea_codes = 1:3)
## Dataframe with factors
ex5 <- data.frame(x = factor(c("Yea", "Nay", "Yea")),
                  y = factor(c("Yea", "Nay", NA)))
y <- response_matrix(ex5, yea_codes = "Yea", nay_codes = "Nay")
is.response_matrix(ex5)
is.response_matrix(y)
as.response_matrix(ex5, yea_codes = "Yea", nay_codes = "Nay")
as.response_matrix(y,   yea_codes = "Yea", nay_codes = "Nay")

}
